+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## TgCore Changelog
 
+### 1.0.92
+* Added: `tg.monitor.uptime_stats()`
+* Added: `tg.monitor.uptime_stats_by_id()`
+* Added: `tg.monitor.get_by_id()`
+* Added: `tg.monitor.reset()`
+* Added: `tg.monitor.pause()`
+* Added: `tg.monitor.start()`
+* Added: `tg.monitor.update()`
+
 ### 1.0.91
 * Added: `tg.monitor.get()`
 * Added: `tg.monitor.create()`